% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{collect_wholesample}
\alias{collect_wholesample}
\title{Collect wholesample}
\usage{
collect_wholesample(db, table)
}
\arguments{
\item{db}{database}

\item{table}{name of the table}
}
\value{
a table
}
\description{
Collect wholesample
}
\examples{

\dontrun{
collect_wholesample(db = database_signauxfaibles, table = "wholesample")
}

}
