% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_samples.R
\name{compute_wholesample_nbdebits}
\alias{compute_wholesample_nbdebits}
\title{Compue wholesample nbdebits}
\usage{
compute_wholesample_nbdebits(db, name, start, end)
}
\arguments{
\item{db}{database}

\item{name}{name of the table}

\item{start}{start date}

\item{end}{end date}
}
\value{
a table in the database
}
\description{
Compue wholesample nbdebits
}
\examples{

\dontrun{
compute_wholesample_nbdebits(
db = database_signauxfaibles,
name = "wholesample_nbdebits",
start = "2013-01-01",
end = "2017-03-01")
}

}
