% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siret.R
\name{get_dettecumulee}
\alias{get_dettecumulee}
\title{Get dettecumulee}
\usage{
get_dettecumulee(db, siret)
}
\arguments{
\item{db}{database name}

\item{siret}{siret number}
}
\value{
a tibble
}
\description{
Get dettecumulee
}
\examples{

\dontrun{
get_dettecumulee(db = database_signauxfaibles, siret = "37850713100027")
}

}
