% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_table_cotisation}
\alias{import_table_cotisation}
\title{Import table cotisation}
\usage{
import_table_cotisation(path)
}
\arguments{
\item{path}{path to the data}
}
\value{
a tibble
}
\description{
Import table cotisation
}
\examples{
\dontrun{
import_table_cotisation(path = "raw-data/urssaf/Urssaf_bourgogne_Cotis_dues_histo_31_08_2016.txt")
}

}
