% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_table_effectif2}
\alias{import_table_effectif2}
\title{Import the new version of table effectif}
\usage{
import_table_effectif2(path)
}
\arguments{
\item{path}{the path to the CSV file}
}
\value{
a tibble
}
\description{
Import the new version of table effectif
}
\examples{
\dontrun{
table_effectif2 <- import_table_effectif2(path = "raw-data/urssaf/Urssaf_emploi_BFC_200401_201609.csv")
}
}
