% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/path_.R
\name{path_}
\alias{path_}
\title{Separate Path}
\usage{
path_(path = NULL, copy2clip = is.global(2))
}
\arguments{
\item{path}{A path to the file/directory.}

\item{copy2clip}{logical.  If \code{TRUE} attempts to copy to the clipboard.
Default copies to the clipboard if used interactively.}
}
\description{
Separate a path into hierarchical drectories and file.
}
\examples{
path_("C:/Users/Tyler/Desktop/guy")
}

