% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siret.R
\name{plot_dettecumulee}
\alias{plot_dettecumulee}
\title{Plot dettecumulee}
\usage{
plot_dettecumulee(db, siret, variable)
}
\arguments{
\item{db}{a database connexion}

\item{siret}{a siret number}

\item{variable}{a variable name ("montant_part_patronale", "dettecumulee", "montant_part_patronale")}
}
\value{
a ggplot
}
\description{
Plot dettecumulee
}
\examples{

}
