% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/path_.R
\name{print.path_}
\alias{print.path_}
\title{Prints a path_ object}
\usage{
\method{print}{path_}(x, ...)
}
\arguments{
\item{x}{A \code{path_} object.}

\item{\ldots}{Other arguments passed to \code{\link[base]{cat}}.}
}
\description{
Prints a path_ object.
}

