% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree_.R
\name{print.tree_}
\alias{print.tree_}
\title{Prints a tree_ object}
\usage{
\method{print}{tree_}(x, ...)
}
\arguments{
\item{x}{A \code{tree_} object.}

\item{\ldots}{Other arguments passed to \code{\link[base]{cat}}.}
}
\description{
Prints a tree_ object.
}

