% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rstudio_.R
\name{rstudio_}
\alias{rstudio_}
\title{Generate a Windows Tree}
\usage{
rstudio_(path = "temp")
}
\arguments{
\item{path}{A path to the directory.}
}
\description{
Given a directory, generate a Windows based tree diagram.
}
\examples{
\dontrun{
rstudio_()
}
}

