% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree_.R
\name{tree_}
\alias{tree_}
\title{Generate a Windows Tree}
\usage{
tree_(path = ".", out = NULL, include.files = TRUE, additional = NULL,
  copy2clip = FALSE)
}
\arguments{
\item{path}{A path to the directory.}

\item{out}{An optional outfle (typically .txt) to print to.}

\item{include.files}{logical.  If \code{TRUE} files and drectories will be
included.}

\item{additional}{Additonal arguments passed to
href{https://www.microsoft.com/resources/documentation/windows/xp/all/proddocs/en-us/tree.mspx?mfr=true}{\code{tree}}.}

\item{copy2clip}{logical.  If \code{TRUE} attempts to copy to the clipboard.}
}
\description{
Given a directory, generate a Windows based tree diagram.
}
\examples{
\dontrun{
tree_()
}
}

